Handle editors in the tracker plugin
